Promoted AzureJobStore's retry logic to general utility
