+++ src/toil/jobStores/azureJobStore.py
@@ -39,6 +39,7 @@ from bd2k.util import strict_bool
 
 from bd2k.util.threading import ExceptionalThread
 
+from toil.jobStores.utils import retry
 from toil.jobWrapper import JobWrapper
 from toil.jobStores.abstractJobStore import (AbstractJobStore, NoSuchJobException,
                                              ConcurrentFileModificationException,
@@ -442,13 +443,13 @@ class AzureJobStore(AbstractJobStore):
 
     def _getOrCreateTable(self, tableName):
         # This will not fail if the table already exists.
-        for attempt in retry_on_error():
+        for attempt in retry_azure():
             with attempt:
                 self.tableService.create_table(tableName)
         return AzureTable(self.tableService, tableName)
 
     def _getOrCreateBlobContainer(self, containerName):
-        for attempt in retry_on_error():
+        for attempt in retry_azure():
             with attempt:
                 self.blobService.create_container(containerName)
         return AzureBlobContainer(self.blobService, containerName)
@@ -621,7 +622,7 @@ class AzureTable(object):
                 if 'PartitionKey' not in kwargs['entity']:
                     kwargs['entity']['PartitionKey'] = self.defaultPartition
 
-            for attempt in retry_on_error():
+            for attempt in retry_azure():
                 with attempt:
                     return function(**kwargs)
 
@@ -693,7 +694,7 @@ class AzureBlobContainer(object):
             function = getattr(self.blobService, name)
             kwargs['container_name'] = self.containerName
 
-            for attempt in retry_on_error():
+            for attempt in retry_azure():
                 with attempt:
                     return function(**kwargs)
 
@@ -752,7 +753,6 @@ class AzureJob(JobWrapper):
         return item
 
 
-
 def defaultRetryPredicate(exception):
     """
     >>> defaultRetryPredicate(socket.error())
@@ -784,100 +784,6 @@ def defaultRetryPredicate(exception):
                 "service unavailable",
                 "server is busy")))
 
-def retry_on_error(delays=(0, 1, 1, 4, 16, 64), timeout=300, predicate=defaultRetryPredicate):
-    """
-    Retry an EC2 operation while the failure matches a given predicate and until a given timeout
-    expires, waiting a given amount of time in between attempts. This function is a generator
-    that yields contextmanagers. See doctests below for example usage.
 
-    :param Iterable[float] delays: an interable yielding the time in seconds to wait before each 
-           retried attempt, the last element of the iterable will be repeated. 
-
-    :param float timeout: a overall timeout that should not be exceeded for all attempts together. 
-           This is a best-effort mechanism only and it won't abort an ongoing attempt, even if the 
-           timeout expires during that attempt. 
-
-    :param Callable[[Exception],bool] predicate: a unary callable returning True if another 
-           attempt should be made to recover from the given exception. 
-
-    :return: a generator yielding context managers, one per attempt
-    :rtype: Iterator
-
-    Retry for a limited amount of time:
-    
-    >>> true = lambda _:True
-    >>> false = lambda _:False
-    >>> i = 0
-    >>> for attempt in retry_on_error( delays=[0], timeout=.1, predicate=true ):
-    ...     with attempt:
-    ...         i += 1
-    ...         raise RuntimeError('foo')
-    Traceback (most recent call last):
-    ...
-    RuntimeError: foo
-    >>> i > 1
-    True
-
-    If timeout is 0, do exactly one attempt:
-    
-    >>> i = 0
-    >>> for attempt in retry_on_error( timeout=0 ):
-    ...     with attempt:
-    ...         i += 1
-    ...         raise RuntimeError( 'foo' )
-    Traceback (most recent call last):
-    ...
-    RuntimeError: foo
-    >>> i
-    1
-
-    Don't retry on success:
-    
-    >>> i = 0
-    >>> for attempt in retry_on_error( delays=[0], timeout=.1, predicate=true ):
-    ...     with attempt:
-    ...         i += 1
-    >>> i
-    1
-
-    Don't retry on unless predicate returns True:
-    
-    >>> i = 0
-    >>> for attempt in retry_on_error( delays=[0], timeout=.1, predicate=false):
-    ...     with attempt:
-    ...         i += 1
-    ...         raise RuntimeError( 'foo' )
-    Traceback (most recent call last):
-    ...
-    RuntimeError: foo
-    >>> i
-    1
-    """
-    if timeout > 0:
-        go = [None]
-
-        @contextmanager
-        def repeated_attempt(delay):
-            try:
-                yield
-            except Exception as e:
-                if time.time() + delay < expiration and predicate(e):
-                    logger.info('Got %s, trying again in %is.', e, delay)
-                    time.sleep(delay)
-                else:
-                    raise
-            else:
-                go.pop()
-
-        delays = iter(delays)
-        expiration = time.time() + timeout
-        delay = next(delays) 
-        while go:
-            yield repeated_attempt(delay)
-            delay = next(delays, delay)
-    else:
-        @contextmanager
-        def single_attempt():
-            yield
-
-        yield single_attempt()
+def retry_azure(delays=(0, 1, 1, 4, 16, 64), timeout=300, predicate=defaultRetryPredicate):
+    retry(delays=delays, timeout=timeout, predicate=predicate)
